* Fixed: Loading of JPEG photo. Image binary data were affected by TStringList. Store them just as a string type. * Fixed: Reset list filter if different file is opened to avoid confusing empty list after file open. * Fixed: Decoding long quoted-printable text due to range check error of small index variable type.

+++ trunk/UQuotedPrintable.pas
@@ -20,7 +20,7 @@ const
 
 function DecodeQuotedPrintable(Text: string): string;
 var
-  O, Count, WS: Byte;
+  O, Count, WS: Integer;
   I: integer;
   InBuf: array[0..Pred(MaxLine)] of Byte;
   OutBuf: array[0..Pred(MaxLine)] of Byte;
